pyproject.toml
[project]
name = "power-market-pipeline"
version = "0.1.0"
description = "Power market data pipeline for ISO electricity markets"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "requests>=2.31.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "pyarrow>=14.0.0",
    "python-dotenv>=1.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation>=0.41b0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "apscheduler>=3.10.0",
    "structlog>=23.2.0",
]

[project.scripts]
pmp = "power_market_pipeline.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "ipython>=8.17.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
